split up App.test.tsx
