+++ src/__mocks__/monaco-editor.ts
@@ -1,10 +1,6 @@
-const monacoEditorMock = {
-  MarkerSeverity: {
-      Hint: 1,
-      Info: 2,
-      Warning: 4,
-      Error: 8,
-  },
+export const MarkerSeverity = {
+  Hint: 1,
+  Info: 2,
+  Warning: 4,
+  Error: 8,
 };
-
-export default monacoEditorMock;
