+++ src/__tests__/language-handler/testAntFiles/grammarAnnotationKeywordTestFiles/encodes.txt
