+++ src/__tests__/language-handler/testAntFiles/grammarAnnotationKeywordTestFiles/hasProperty.txt
