+++ src/__tests__/language-handler/testAntFiles/grammarAnnotationKeywordTestFiles/hasVersion.txt
