+++ src/__tests__/language-handler/testAntFiles/grammarAnnotationKeywordTestFiles/isPropertyOf.txt
