+++ src/__tests__/language-handler/testAntFiles/predefConstantNoError.ant
