+++ src/__tests__/language-handler/testAntFiles/predefConstantTypeAssignError.ant
