+++ src/__tests__/language-handler/testAntFiles/predefConstantValueAssignError.ant
