+++ src/__tests__/language-handler/testAntFiles/reactionCompartment.ant
