+++ src/__tests__/language-handler/testAntFiles/reactionNoRateRule.ant
