+++ src/index.tsx
@@ -2,8 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './testing/reportWebVitals';
-// index.ts
 
 import { customLanguageWorker } from './language-handler/antlr/CustomLanguageWorker';
